[FEAT] NFT판매 등록 컨트랙트, 판매 종료 API 연동 (#S08P22D208-412, #S08P22D208-414)

+++ Android/app/src/main/java/com/ssafy/kkaddak/domain/entity/market/NftDetailItem.kt
@@ -1,12 +1,14 @@
 package com.ssafy.kkaddak.domain.entity.market
 
+import java.math.BigInteger
+
 data class NftDetailItem(
     val nftImageUrl: String,
     val creatorNickname: String,
     val songTitle: String,
     val saleHistoryList: List<HistoryItem>,
     val isSelling: Boolean,
-    val price: Double,
+    val price: BigInteger,
     var isLike: Boolean,
     val sellerNickname: String,
     val sellerAccount: String
